Got rid of configurable Added tests for helper functions (has, get, set) of Configuration
