+++ lib/cluster_process.js
@@ -2,8 +2,10 @@ var cluster = require('cluster'),
     path = require('path'),
     RPC = require('./rpc'),
     RPCCallback = require('./rpc-callback'),
-    Configurable = require('./configurable'),
+    Configuration = require('./configuration'),
+    EventEmitterEx = require('./event_emitter_ex'),
     LusterClusterProcessError = require('./errors').LusterClusterProcessError,
+    LusterConfigurationError = require('./errors').LusterConfigurationError,
     ClusterProcess;
 
 /**
@@ -23,9 +25,9 @@ function has(context, propName) {
 /**
  * @constructor
  * @class ClusterProcess
- * @augments Configurable
+ * @augments EventEmitterEx
  */
-ClusterProcess = Configurable.create(function ClusterProcess() {
+ClusterProcess = EventEmitterEx.create(function ClusterProcess() {
     ClusterProcess.__super.apply(this, arguments);
 
     /** @private */
@@ -34,6 +36,12 @@ ClusterProcess = Configurable.create(function ClusterProcess() {
     this._initialized = false;
     /** @private */
     this.extensions = {};
+
+    /**
+     * @type {Configuration}
+     * @public
+     */
+    this.config = null;
 
     this.once('configured', this._onConfigured.bind(this));
 
@@ -65,6 +73,46 @@ Object.defineProperty(ClusterProcess.prototype, 'isWorker', {
 });
 
 /**
+ * @event ClusterProcess#configured
+ */
+
+/**
+ * @fires ClusterProcess#configured
+ * @param {Object} config
+ * @param {Boolean} [applyEnv=true]
+ * @param {String} [basedir=process.cwd()] for Configuration#resolve relative paths
+ * @returns {ClusterProcess} this
+ * @throws {LusterConfigurationError} if configuration check failed (check errors will be logged to STDERR)
+ * @public
+ */
+ClusterProcess.prototype.configure = function(config, applyEnv, basedir) {
+    if (typeof applyEnv === 'undefined' || applyEnv) {
+        // @todo may be create a deep copy and do not modify origin?
+        Configuration.applyEnvironment(config);
+    }
+
+    if (typeof(basedir) === 'undefined') {
+        basedir = process.cwd();
+    }
+
+    if (Configuration.check(config) > 0) {
+        this.emit('error',
+            LusterConfigurationError.createError(
+                LusterConfigurationError.CODES.CONFIGURATION_CHECK_FAILED));
+    } else {
+        this.config = Configuration.extend(config, basedir);
+
+        // hack to tweak underlying EventEmitter max listeners
+        // if your luster-based app extensively use luster events
+        this.setMaxListeners(this.config.get('maxEventListeners', 100));
+
+        this.emit('configured');
+    }
+
+    return this;
+};
+
+/**
  * @param {String} name
  * @param {Function} callback function(error)
  */
@@ -78,7 +126,7 @@ ClusterProcess.prototype.loadExtension = function(name, callback) {
 
     // if `config` was an Object then it became instance of Configuration
     // else returns original value
-    config = this.extendConfig(config);
+    config = Configuration.extend(config, this.config.getBaseDir());
 
     if (extension.configure.length > 2) {
         setImmediate(function() {
